Improving Newsletter, Adding Products, Adding email upon registration

+++ danko-handmade/src/main/java/com/danko/danko_handmade/product/model/Product.java
@@ -54,5 +54,5 @@ public class Product {
     @Column(nullable = false)
     private boolean active;
 
-    private Long itemsSold;
+    private int itemsSold;
 }
